Make the unit test for compute_e check with concrete value and moved to independent file for to share among bitcoin_schnorr and zilliqa_schnorr.
Essentialy how to compute e value is depends on each specification of platforms, but current implemented specification is same. For now it can be shared.
