+++ Cargo.toml
@@ -20,5 +20,4 @@ tag = "v0.2.3"
 
 [dev-dependencies]
 hex = "0.3.2"
-sha2 = "0.8.1"
